CR: Break out the SettingsScreen and HomeScreen into their own files
